Moved components to their own file App.jsx file was getting a little unruly
